refactor: remove infra module
